+++ services/web/server/src/simcore_service_webserver/projects/projects_api.py
@@ -1243,7 +1243,7 @@ async def try_open_project_for_user(
             project_uuid,
             ProjectStatus.OPENING,
             user_id,
-            await get_user_fullname(app, user_id),
+            await get_user_fullname(app, user_id=user_id),
             notify_users=False,
         ):
             with managed_resource(user_id, client_session_id, app) as user_session:
@@ -1413,22 +1413,23 @@ async def _get_project_lock_state(
         f"{set_user_ids=}",
     )
     usernames: list[FullNameDict] = [
-        await get_user_fullname(app, uid) for uid in set_user_ids
+        await get_user_fullname(app, user_id=uid) for uid in set_user_ids
     ]
     # let's check if the project is opened by the same user, maybe already opened or closed in a orphaned session
-    if set_user_ids.issubset({user_id}):
-        if not await _user_has_another_client_open(user_session_id_list, app):
-            # in this case the project is re-openable by the same user until it gets closed
-            log.debug(
-                "project [%s] is in use by the same user [%s] that is currently disconnected, so it is unlocked for this specific user and opened",
-                f"{project_uuid=}",
-                f"{set_user_ids=}",
-            )
-            return ProjectLocked(
-                value=False,
-                owner=Owner(user_id=next(iter(set_user_ids)), **usernames[0]),
-                status=ProjectStatus.OPENED,
-            )
+    if set_user_ids.issubset({user_id}) and not await _user_has_another_client_open(
+        user_session_id_list, app
+    ):
+        # in this case the project is re-openable by the same user until it gets closed
+        log.debug(
+            "project [%s] is in use by the same user [%s] that is currently disconnected, so it is unlocked for this specific user and opened",
+            f"{project_uuid=}",
+            f"{set_user_ids=}",
+        )
+        return ProjectLocked(
+            value=False,
+            owner=Owner(user_id=next(iter(set_user_ids)), **usernames[0]),
+            status=ProjectStatus.OPENED,
+        )
     # the project is opened in another tab or browser, or by another user, both case resolves to the project being locked, and opened
     log.debug(
         "project [%s] is in use by another user [%s], so it is locked",
@@ -1712,7 +1713,9 @@ async def remove_project_dynamic_services(
         user_id,
     )
 
-    user_name_data: FullNameDict = user_name or await get_user_fullname(app, user_id)
+    user_name_data: FullNameDict = user_name or await get_user_fullname(
+        app, user_id=user_id
+    )
 
     user_role: UserRole | None = None
     try:
