+++ services/web/server/src/simcore_service_webserver/users/_users_repository.py
@@ -1,9 +1,14 @@
 import contextlib
+from typing import Any
 
+import simcore_postgres_database.errors as db_errors
 import sqlalchemy as sa
 from aiohttp import web
+from common_library.groups_enums import GroupType
+from common_library.users_enums import UserRole
 from models_library.groups import GroupID
-from models_library.users import UserBillingDetails, UserID, UserPermission
+from models_library.users import UserBillingDetails, UserID, UserNameID, UserPermission
+from pydantic import TypeAdapter, ValidationError
 from simcore_postgres_database.models.groups import groups, user_to_groups
 from simcore_postgres_database.models.products import products
 from simcore_postgres_database.models.users import UserStatus, users
@@ -18,15 +23,29 @@ from simcore_postgres_database.utils_repos import (
     pass_or_acquire_connection,
     transaction_context,
 )
-from simcore_postgres_database.utils_users import UsersRepo
-from simcore_service_webserver.users.exceptions import UserNotFoundError
+from simcore_postgres_database.utils_users import (
+    UsersRepo,
+    generate_alternative_username,
+)
 from sqlalchemy.engine.row import Row
 from sqlalchemy.ext.asyncio import AsyncConnection, AsyncEngine
 
 from ..db.plugin import get_asyncpg_engine
-from .exceptions import BillingDetailsNotFoundError
+from ._common.models import FullNameDict, ToUserUpdateDB
+from .exceptions import (
+    BillingDetailsNotFoundError,
+    UserNameDuplicateError,
+    UserNotFoundError,
+)
 
 _ALL = None
+
+
+def _parse_as_user(user_id: Any) -> UserID:
+    try:
+        return TypeAdapter(UserID).validate_python(user_id)
+    except ValidationError as err:
+        raise UserNotFoundError(uid=user_id, user_id=user_id) from err
 
 
 async def get_user_or_raise(
@@ -35,7 +54,7 @@ async def get_user_or_raise(
     *,
     user_id: UserID,
     return_column_names: list[str] | None = _ALL,
-) -> Row:
+) -> dict[str, Any]:
     if return_column_names == _ALL:
         return_column_names = list(users.columns.keys())
 
@@ -51,7 +70,8 @@ async def get_user_or_raise(
         row = await result.first()
         if row is None:
             raise UserNotFoundError(uid=user_id)
-        return row
+        user: dict[str, Any] = row._asdict()
+        return user
 
 
 async def get_users_ids_in_group(
@@ -65,6 +85,65 @@ async def get_users_ids_in_group(
             sa.select(user_to_groups.c.uid).where(user_to_groups.c.gid == group_id)
         )
         return {row.uid async for row in result}
+
+
+async def get_user_id_from_pgid(app: web.Application, primary_gid: int) -> UserID:
+    async with pass_or_acquire_connection(engine=get_asyncpg_engine(app)) as conn:
+        user_id: UserID = await conn.scalar(
+            sa.select(users.c.id).where(users.c.primary_gid == primary_gid)
+        )
+        return user_id
+
+
+async def get_user_fullname(app: web.Application, *, user_id: UserID) -> FullNameDict:
+    """
+    :raises UserNotFoundError:
+    """
+    user_id = _parse_as_user(user_id)
+
+    async with pass_or_acquire_connection(engine=get_asyncpg_engine(app)) as conn:
+        result = await conn.stream(
+            sa.select(
+                users.c.first_name,
+                users.c.last_name,
+            ).where(users.c.id == user_id)
+        )
+        user = await result.first()
+        if not user:
+            raise UserNotFoundError(uid=user_id)
+
+        return FullNameDict(
+            first_name=user.first_name,
+            last_name=user.last_name,
+        )
+
+
+async def get_guest_user_ids_and_names(
+    app: web.Application,
+) -> list[tuple[UserID, UserNameID]]:
+    async with pass_or_acquire_connection(engine=get_asyncpg_engine(app)) as conn:
+        result = await conn.stream(
+            sa.select(users.c.id, users.c.name).where(users.c.role == UserRole.GUEST)
+        )
+
+        return TypeAdapter(list[tuple[UserID, UserNameID]]).validate_python(
+            [(row.id, row.name) async for row in result]
+        )
+
+
+async def get_user_role(app: web.Application, *, user_id: UserID) -> UserRole:
+    """
+    :raises UserNotFoundError:
+    """
+    user_id = _parse_as_user(user_id)
+
+    async with pass_or_acquire_connection(engine=get_asyncpg_engine(app)) as conn:
+        user_role = await conn.scalar(
+            sa.select(users.c.role).where(users.c.id == user_id)
+        )
+        if user_role is None:
+            raise UserNotFoundError(uid=user_id)
+        return UserRole(user_role)
 
 
 async def list_user_permissions(
@@ -247,3 +326,150 @@ async def get_user_billing_details(
         if not row:
             raise BillingDetailsNotFoundError(user_id=user_id)
         return UserBillingDetails.model_validate(row)
+
+
+#
+# USER PROFILE
+#
+
+
+_GROUPS_SCHEMA_TO_DB = {
+    "gid": "gid",
+    "label": "name",
+    "description": "description",
+    "thumbnail": "thumbnail",
+    "accessRights": "access_rights",
+}
+
+
+def _convert_groups_db_to_schema(
+    db_row: Row, *, prefix: str | None = "", **kwargs
+) -> dict:
+    # NOTE: Deprecated. has to be replaced with
+    converted_dict = {
+        k: db_row[f"{prefix}{v}"]
+        for k, v in _GROUPS_SCHEMA_TO_DB.items()
+        if f"{prefix}{v}" in db_row
+    }
+    converted_dict.update(**kwargs)
+    converted_dict["inclusionRules"] = {}
+    return converted_dict
+
+
+async def get_user_profile(app: web.Application, *, user_id: UserID) -> dict[str, Any]:
+
+    user_profile: dict[str, Any] = {}
+    user_primary_group = everyone_group = {}
+    user_standard_groups = []
+    user_id = _parse_as_user(user_id)
+
+    async with pass_or_acquire_connection(engine=get_asyncpg_engine(app)) as conn:
+        result = await conn.stream(
+            sa.select(users, groups, user_to_groups.c.access_rights)
+            .select_from(
+                users.join(user_to_groups, users.c.id == user_to_groups.c.uid).join(
+                    groups, user_to_groups.c.gid == groups.c.gid
+                )
+            )
+            .where(users.c.id == user_id)
+            .order_by(sa.asc(groups.c.name))
+            .set_label_style(sa.LABEL_STYLE_TABLENAME_PLUS_COL)
+        )
+
+        async for row in result:
+            if not user_profile:
+                user_profile = {
+                    "id": row.users_id,
+                    "user_name": row.users_name,
+                    "first_name": row.users_first_name,
+                    "last_name": row.users_last_name,
+                    "login": row.users_email,
+                    "role": row.users_role,
+                    "privacy_hide_fullname": row.users_privacy_hide_fullname,
+                    "privacy_hide_email": row.users_privacy_hide_email,
+                    "expiration_date": (
+                        row.users_expires_at.date() if row.users_expires_at else None
+                    ),
+                }
+                assert user_profile["id"] == user_id  # nosec
+
+            if row.groups_type == GroupType.EVERYONE:
+                everyone_group = _convert_groups_db_to_schema(
+                    row,
+                    prefix="groups_",
+                    accessRights=row["user_to_groups_access_rights"],
+                )
+            elif row.groups_type == GroupType.PRIMARY:
+                user_primary_group = _convert_groups_db_to_schema(
+                    row,
+                    prefix="groups_",
+                    accessRights=row["user_to_groups_access_rights"],
+                )
+            else:
+                user_standard_groups.append(
+                    _convert_groups_db_to_schema(
+                        row,
+                        prefix="groups_",
+                        accessRights=row["user_to_groups_access_rights"],
+                    )
+                )
+
+    if not user_profile:
+        raise UserNotFoundError(uid=user_id)
+
+    # NOTE: expirationDate null is not handled properly in front-end.
+    # https://github.com/ITISFoundation/osparc-simcore/issues/5244
+    optional = {}
+    if user_profile.get("expiration_date"):
+        optional["expiration_date"] = user_profile["expiration_date"]
+
+    return dict(
+        id=user_profile["id"],
+        user_name=user_profile["user_name"],
+        first_name=user_profile["first_name"],
+        last_name=user_profile["last_name"],
+        login=user_profile["login"],
+        role=user_profile["role"],
+        groups={
+            "me": user_primary_group,
+            "organizations": user_standard_groups,
+            "all": everyone_group,
+        },
+        privacy={
+            "hide_fullname": user_profile["privacy_hide_fullname"],
+            "hide_email": user_profile["privacy_hide_email"],
+        },
+        **optional,
+    )
+
+
+async def update_user_profile(
+    app: web.Application,
+    *,
+    user_id: UserID,
+    update: ToUserUpdateDB,
+) -> None:
+    """
+    Raises:
+        UserNotFoundError
+        UserNameAlreadyExistsError
+    """
+    user_id = _parse_as_user(user_id)
+
+    if updated_values := update.to_db():
+
+        async with transaction_context(engine=get_asyncpg_engine(app)) as conn:
+            query = users.update().where(users.c.id == user_id).values(**updated_values)
+
+            try:
+                await conn.execute(query)
+
+            except db_errors.UniqueViolation as err:
+                user_name = updated_values.get("name")
+
+                raise UserNameDuplicateError(
+                    user_name=user_name,
+                    alternative_user_name=generate_alternative_username(user_name),
+                    user_id=user_id,
+                    updated_values=updated_values,
+                ) from err
